LIC_TCM/requirements.txt
# TCM (CVPR 2023) 复现项目依赖
# PyTorch 核心 (根据你的 CUDA 版本选择)
torch>=2.0.0
torchvision>=0.15.0

# 图像压缩核心库
compressai>=1.2.0

# 模型组件
timm>=0.9.0

# 数据处理
scipy>=1.10.0
numpy>=1.24.0
pillow>=9.5.0
pandas>=2.0.0

# 可视化
matplotlib>=3.7.0
tensorboard>=2.13.0

# 训练辅助
tqdm>=4.65.0
pyyaml>=6.0
ninja>=1.11.0  # 可选，加速编译

# 评估指标
opencv-python>=4.7.0
scikit-image>=0.20.0
